+++ src/Sidebar/Logo/styled.js
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+
+export const Wrapper = styled.div`
+  width: 77px;
+  height: 52px;
+  padding-top: 21px;
+  padding-bottom: 21px;
+  padding-left: 9px;
+  padding-right: 9px;
+  margin: 0 auto;
+`;
